Moved HD > CooldownPlanner in QE Modules. Renamed HD to CooldownPlanner.
