--- src/Modules/HolyDiverModules/Functions/IconFunctions/TalentIcons.js
@@ -1,683 +1,0 @@
-import React from "react";
-// Paladin Icons
-import AvengingCrusaderIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/AvengingCrusaderIcon.jpg";
-import AwakeningIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/AwakeningIcon.jpg";
-import BeaconOfFaithIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/BeaconOfFaithIcon.jpg";
-import BeaconOfVirtueIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/BeaconOfVirtueIcon.jpg";
-import BestowFaithIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/BestowFaithIcon.jpg";
-import BlindingLightIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/BlindingLightIcon.jpg";
-import CavalierIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/CavalierIcon.jpg";
-import CrusadersMightIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/CrusadersMightIcon.jpg";
-import DivinePurposeIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/DivinePurposeIcon.jpg";
-import FistOfJusticeIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/FistOfJusticeIcon.jpg";
-import GlimmerOfLightIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/GlimmerOfLightIcon.jpg";
-import HolyAvengerIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/HolyAvengerIcon.jpg";
-import HolyPrismIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/HolyPrismIcon.jpg";
-import JudgmentOfLightIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/JudgmentOfLightIcon.jpg";
-import LightsHammerIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/LightsHammerIcon.jpg";
-import RepentanceIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/RepentanceIcon.jpg";
-import RuleOfLawIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/RuleOfLawIcon.jpg";
-import SavedByTheLightIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/SavedByTheLightIcon.jpg";
-import SeraphimIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/SeraphimIcon.jpg";
-import UnbreakableSpiritIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/UnbreakableSpiritIcon.jpg";
-import SanctifiedWrathIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/Talents/SanctifiedWrathIcon.jpg";
-
-// Shaman Icons
-import AncestralProtectionTotemIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/AncestralProtectionTotemIcon.jpg";
-import AncestralVigorIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/AncestralVigorIcon.jpg";
-import AscendanceIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/AscendanceIcon.jpg";
-import CloudburstTotemIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/CloudburstTotemIcon.jpg";
-import DelugeIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/DelugeIcon.jpg";
-import DownpourIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/DownpourIcon.jpg";
-import EarthenWallTotemIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/EarthenWallTotemIcon.jpg";
-import EarthgrabTotemIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/EarthgrabTotemIcon.jpg";
-import EchoOfTheElementsIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/EchoOfTheElementsIcon.jpg";
-import FlashFloodIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/FlashFloodIcon.jpg";
-import GracefulSpiritIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/GracefulSpiritIcon.jpg";
-import HighTideIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/HighTideIcon.jpg";
-import NaturesGuardianIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/NaturesGuardianIcon.jpg";
-import SpiritWolfIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/SpiritWolfIcon.jpg";
-import StaticChargeIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/StaticChargeIcon.jpg";
-import SurgeOfEarthIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/SurgeOfEarthIcon.jpg";
-import TorrentIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/TorrentIcon.jpg";
-import UndulationIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/UndulationIcon.jpg";
-import UnleashLifeIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/UnleashLifeIcon.jpg";
-import WellspringIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/WellspringIcon.jpg";
-import WindRushTotemIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/Talents/WindRushTotemIcon.jpg";
-
-// Druid Icons
-import AbundanceIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/AbundanceIcon.jpg";
-import BalanceAffinityIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/BalanceAffinityIcon.jpg";
-import CenarionWardIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/CenarionWardIcon.jpg";
-import CultivationIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/CultivationIcon.jpg";
-import FeralAffinityIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/FeralAffinityIcon.jpg";
-import FlourishIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/FlourishIcon.jpg";
-import GerminationIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/GerminationIcon.jpg";
-import GuardianAffinityIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/GuardianAffinityIcon.jpg";
-import HeartOfTheWildIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/HeartOfTheWildIcon.jpg";
-import IncarnationTreeOfLifeIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/IncarnationTreeOfLifeIcon.jpg";
-import InnerPeaceIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/InnerPeaceIcon.jpg";
-import MassEntanglementIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/MassEntanglementIcon.jpg";
-import MightyBashIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/MightyBashIcon.jpg";
-import NourishIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/NourishIcon.jpg";
-import OvergrowthIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/OvergrowthIcon.jpg";
-import PhotosynthesisIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/PhotosynthesisIcon.jpg";
-import RenewalIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/RenewalIcon.jpg";
-import SoulOfTheForestIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/SoulOfTheForestIcon.jpg";
-import SpringBlossomsIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/SpringBlossomsIcon.jpg";
-import TigerDashIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/TigerDashIcon.jpg";
-import WildChargeIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Talents/WildChargeIcon.jpg";
-
-// Monk Icons
-import CelerityIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/CelerityIcon.jpg";
-import ChiBurstIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/ChiBurstIcon.jpg";
-import ChiTorpedoIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/ChiTorpedoIcon.jpg";
-import ChiWaveIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/ChiWaveIcon.jpg";
-import DampenHarmIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/DampenHarmIcon.jpg";
-import DiffuseMagicIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/DiffuseMagicIcon.jpg";
-import FocusedThunderIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/FocusedThunderIcon.jpg";
-import HealingElixirIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/HealingElixirIcon.jpg";
-import InvokeChiJiTheRedCraneIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/InvokeChiJiTheRedCraneIcon.jpg";
-import LifecyclesIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/LifecyclesIcon.jpg";
-import ManaTeaIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/ManaTeaIcon.jpg";
-import MistWrapIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/MistWrapIcon.jpg";
-import RefreshingJadeWindIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/RefreshingJadeWindIcon.jpg";
-import RingOfPeaceIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/RingOfPeaceIcon.jpg";
-import RisingMistIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/RisingMistIcon.jpg";
-import SongOfChiJiIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/SongOfChiJiIcon.jpg";
-import SpiritOfTheCraneIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/SpiritOfTheCraneIcon.jpg";
-import SummonJadeSerpentStatueIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/SummonJadeSerpentStatueIcon.jpg";
-import TigersLustIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/TigersLustIcon.jpg";
-import TigerTailSweepIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/TigerTailSweepIcon.jpg";
-import UpwellingIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Talents/UpwellingIcon.jpg";
-
-// Priest - Discipline Icons
-import AngelicFeatherIconDisc from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/AngelicFeatherIcon.jpg";
-import BodyAndSoulIconDisc from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/BodyAndSoulIcon.jpg";
-import CastigationIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/CastigationIcon.jpg";
-import ContritionIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/ContritionIcon.jpg";
-import DivineStarIconDisc from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/DivineStarIcon.jpg";
-import DominantMindIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/DominantMindIcon.jpg";
-import EvangelismIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/EvangelismIcon.jpg";
-import HaloIconDisc from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/HaloIcon.jpg";
-import LenienceIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/LenienceIcon.jpg";
-import MasochismIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/MasochismIcon.jpg";
-import MindbenderIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/MindbenderIcon.jpg";
-import PowerWordSolaceIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/PowerWordSolaceIcon.jpg";
-import PsychicVoiceIconDisc from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/PsychicVoiceIcon.jpg";
-import PurgeTheWickedIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/PurgeTheWickedIcon.jpg";
-import SchismIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/SchismIcon.jpg";
-import ShadowCovenantIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/ShadowCovenantIcon.jpg";
-import ShieldDisciplineIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/ShieldDisciplineIcon.jpg";
-import ShiningForceIconDisc from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/ShiningForceIcon.jpg";
-import SinsOfTheManyIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/SinsOfTheManyIcon.jpg";
-import SpiritShellIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/SpiritShellIcon.jpg";
-import TwistOfFateIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Talents/TwistOfFateIcon.jpg";
-
-// Priest - Holy Icons
-import AfterlifeIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/AfterlifeIcon.jpg";
-import AngelicFeatherIconHoly from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/AngelicFeatherIcon.jpg";
-import AngelsMercyIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/AngelsMercyIcon.jpg";
-import ApotheosisIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/ApotheosisIcon.jpg";
-import BenedictionIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/BenedictionIcon.jpg";
-import BindingHealIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/BindingHealIcon.jpg";
-import BodyAndSoulIconHoly from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/BodyAndSoulIcon.jpg";
-import CensureIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/CensureIcon.jpg";
-import CosmicRippleIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/CosmicRippleIcon.jpg";
-import DivineStarIconHoly from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/DivineStarIcon.jpg";
-import EnlightenmentIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/EnlightenmentIcon.jpg";
-import GuardianAngelIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/GuardianAngelIcon.jpg";
-import HaloIconHoly from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/HaloIcon.jpg";
-import HolyWordSalvationIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/HolyWordSalvationIcon.jpg";
-import LightOfTheNaaruIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/LightOfTheNaaruIcon.jpg";
-import PrayerCircleIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/PrayerCircleIcon.jpg";
-import PsychicVoiceIconHoly from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/PsychicVoiceIcon.jpg";
-import RenewedFaithIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/RenewedFaithIcon.jpg";
-import ShiningForceIconHoly from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/ShiningForceIcon.jpg";
-import SurgeOfLightIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/SurgeOfLightIcon.jpg";
-import TrailOfLightIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Talents/TrailOfLightIcon.jpg";
-
-
-
-
-
-
-
-
-export default function talentIcons(props) {
-  let source = "";
-  let alt = "";
-
-  // Paladin Talents
-  if (props === 196926) {
-    source = CrusadersMightIcon;
-    alt = "Crusader's Might";
-  }
-  if (props === 223306) {
-    source = BestowFaithIcon;
-    alt = "Bestow Faith";
-  }
-  if (props === 114158) {
-    source = LightsHammerIcon;
-    alt = "Light's Hammer";
-  }
-  if (props === 157047) {
-    source = SavedByTheLightIcon;
-    alt = "Saved by the Light";
-  }
-  if (props === 183778) {
-    source = JudgmentOfLightIcon;
-    alt = "Judgment of Light";
-  }
-  if (props === 114165) {
-    source = HolyPrismIcon;
-    alt = "Holy Prism";
-  }
-  if (props === 234299) {
-    source = FistOfJusticeIcon;
-    alt = "Fist of Justice";
-  }
-  if (props === 20066) {
-    source = RepentanceIcon;
-    alt = "Repentance";
-  }
-  if (props === 115750) {
-    source = BlindingLightIcon;
-    alt = "Blinding Light";
-  }
-  if (props === 114154) {
-    source = UnbreakableSpiritIcon;
-    alt = "Unbreakable Spirit";
-  }
-  if (props === 230332) {
-    source = CavalierIcon;
-    alt = "Cavalier";
-  }
-  if (props === 214202) {
-    source = RuleOfLawIcon;
-    alt = "Rule of Law";
-  }
-  if (props === 223817) {
-    source = DivinePurposeIcon;
-    alt = "Divine Purpose";
-  }
-  if (props === 105809) {
-    source = HolyAvengerIcon;
-    alt = "Holy Avenger";
-  }
-  if (props === 152262) {
-    source = SeraphimIcon;
-    alt = "Seraphim";
-  }
-  if (props === 53376) {
-    source = SanctifiedWrathIcon;
-    alt = "Sanctified Wrath";
-  }
-  if (props === 216331) {
-    source = AvengingCrusaderIcon;
-    alt = "Avenging Crusader";
-  }
-  if (props === 248033) {
-    source = AwakeningIcon;
-    alt = "Awakening";
-  }
-  if (props === 325966) {
-    source = GlimmerOfLightIcon;
-    alt = "Glimmer of Light";
-  }
-  if (props === 156910) {
-    source = BeaconOfFaithIcon;
-    alt = "Beacon of Faith";
-  }
-  if (props === 200025) {
-    source = BeaconOfVirtueIcon;
-    alt = "Beacon of Virtue";
-  }
-  // Shaman
-  //
-  if (props === 200072) {
-    source = TorrentIcon;
-    alt = "Torrent";
-  }
-  if (props === 200071) {
-    source = UndulationIcon;
-    alt = "Undulation";
-  }
-  if (props === 73685) {
-    source = UnleashLifeIcon;
-    alt = "Unleash Life";
-  }
-  if (props === 108283) {
-    source = EchoOfTheElementsIcon;
-    alt = "Echo of the Elements";
-  }
-  if (props === 200076) {
-    source = DelugeIcon;
-    alt = "Deluge";
-  }
-  if (props === 320746) {
-    source = SurgeOfEarthIcon;
-    alt = "Surge of Earth";
-  }
-  if (props === 260878) {
-    source = SpiritWolfIcon;
-    alt = "Spirit Wolf";
-  }
-  if (props === 51485) {
-    source = EarthgrabTotemIcon;
-    alt = "Earthgrab Totem";
-  }
-  if (props === 265046) {
-    source = StaticChargeIcon;
-    alt = "Static Charge";
-  }
-  if (props === 207401) {
-    source = AncestralVigorIcon;
-    alt = "Ancestral Vigor";
-  }
-  if (props === 198838) {
-    source = EarthenWallTotemIcon;
-    alt = "Earthen Wall Totem";
-  }
-  if (props === 207399) {
-    source = AncestralProtectionTotemIcon;
-    alt = "Ancestral Protection Totem";
-  }
-  if (props === 30884) {
-    source = NaturesGuardianIcon;
-    alt = "Nature's Guardian";
-  }
-  if (props === 192088) {
-    source = GracefulSpiritIcon;
-    alt = "Graceful Spirit";
-  }
-  if (props === 192077) {
-    source = WindRushTotemIcon;
-    alt = "Wind Rush Totem";
-  }
-  if (props === 280614) {
-    source = FlashFloodIcon;
-    alt = "Flash Flood";
-  }
-  if (props === 207778) {
-    source = DownpourIcon;
-    alt = "Downpour";
-  }
-  if (props === 157153) {
-    source = CloudburstTotemIcon;
-    alt = "Cloudburst Totem";
-  }
-  if (props === 157154) {
-    source = HighTideIcon;
-    alt = "High Tide";
-  }
-  if (props === 197995) {
-    source = WellspringIcon;
-    alt = "Wellspring";
-  }
-  if (props === 114052) {
-    source = AscendanceIcon;
-    alt = "Ascendance";
-  }
-  // Druid
-  if (props === 207383) {
-    source = AbundanceIcon;
-    alt = "Abundance";
-  }
-  if (props === 50464) {
-    source = NourishIcon;
-    alt = "Nourish";
-  }
-  if (props === 102351) {
-    source = CenarionWardIcon;
-    alt = "Cenarion Ward";
-  }
-  if (props === 252216) {
-    source = TigerDashIcon;
-    alt = "Tiger Dash";
-  }
-  if (props === 108238) {
-    source = RenewalIcon;
-    alt = "Renewal";
-  }
-  if (props === 132302 || props === 102401) {
-    source = WildChargeIcon;
-    alt = "Wild Charge";
-  }
-  if (props === 197632) {
-    source = BalanceAffinityIcon;
-    alt = "Balance Affinity";
-  }
-  if (props === 197490) {
-    source = FeralAffinityIcon;
-    alt = "Feral Affinity";
-  }
-  if (props === 197491) {
-    source = GuardianAffinityIcon;
-    alt = "Guardian Affinity";
-  }
-  if (props === 5211) {
-    source = MightyBashIcon;
-    alt = "Mighty Bash";
-  }
-  if (props === 102359) {
-    source = MassEntanglementIcon;
-    alt = "Mass Entanglement";
-  }
-  if (props === 319454) {
-    source = HeartOfTheWildIcon;
-    alt = "Heart of the Wild";
-  }
-  if (props === 158478) {
-    source = SoulOfTheForestIcon;
-    alt = "Soul of the Forest";
-  }
-  if (props === 200390) {
-    source = CultivationIcon;
-    alt = "Cultivation";
-  }
-  if (props === 33891) {
-    source = IncarnationTreeOfLifeIcon;
-    alt = "Incarnation: Tree of Life";
-  }
-  if (props === 197073) {
-    source = InnerPeaceIcon;
-    alt = "Inner Peace";
-  }
-  if (props === 207385) {
-    source = SpringBlossomsIcon;
-    alt = "Spring Blossoms";
-  }
-  if (props === 203651) {
-    source = OvergrowthIcon;
-    alt = "Overgrowth";
-  }
-  if (props === 274902) {
-    source = PhotosynthesisIcon;
-    alt = "Photosynthesis";
-  }
-  if (props === 155675) {
-    source = GerminationIcon;
-    alt = "Germination";
-  }
-  if (props === 197721) {
-    source = FlourishIcon;
-    alt = "Flourish";
-  }
-  // Monk
-  if (props === 197900) {
-    source = MistWrapIcon;
-    alt = "Mist Wrap";
-  }
-  if (props === 115098) {
-    source = ChiWaveIcon;
-    alt = "Chi Wave";
-  }
-  if (props === 123986) {
-    source = ChiBurstIcon;
-    alt = "Chi Burst";
-  }
-  if (props === 115173) {
-    source = CelerityIcon;
-    alt = "Celerity";
-  }
-  if (props === 115008) {
-    source = ChiTorpedoIcon;
-    alt = "Chi Torpedo";
-  }
-  if (props === 116841) {
-    source = TigersLustIcon;
-    alt = "Tiger's Lust";
-  }
-  if (props === 197915) {
-    source = LifecyclesIcon;
-    alt = "Lifecycles";
-  }
-  if (props === 210802) {
-    source = SpiritOfTheCraneIcon;
-    alt = "Spirit of the Crane";
-  }
-  if (props === 197908) {
-    source = ManaTeaIcon;
-    alt = "Mana Tea";
-  }
-  if (props === 264348) {
-    source = TigerTailSweepIcon;
-    alt = "Tiger Tail Sweep";
-  }
-  if (props === 198898) {
-    source = SongOfChiJiIcon;
-    alt = "Song of Chi-Ji";
-  }
-  if (props === 116844) {
-    source = RingOfPeaceIcon;
-    alt = "Ring of Peace";
-  }
-  if (props === 122281) {
-    source = HealingElixirIcon;
-    alt = "Healing Elixir";
-  }
-  if (props === 122783) {
-    source = DiffuseMagicIcon;
-    alt = "Diffuse Magic";
-  }
-  if (props === 122278) {
-    source = DampenHarmIcon;
-    alt = "Dampen Harm";
-  }
-  if (props === 115313) {
-    source = SummonJadeSerpentStatueIcon;
-    alt = "Summon Jade Serpent Statue";
-  }
-  if (props === 196725) {
-    source = RefreshingJadeWindIcon;
-    alt = "Refreshing Jade Wind";
-  }
-  if (props === 325197) {
-    source = InvokeChiJiTheRedCraneIcon;
-    alt = "Invoke Chi-Ji, the Red Crane";
-  }
-  if (props === 197895) {
-    source = FocusedThunderIcon;
-    alt = "Focused Thunder";
-  }
-  if (props === 274963) {
-    source = UpwellingIcon;
-    alt = "Upwelling";
-  }
-  if (props === 274909) {
-    source = RisingMistIcon;
-    alt = "Rising Mist";
-  }
-
-  // Priest - Discipline
-  
-  if (props === 197900) {
-    source = CastigationIcon;
-    alt = "Castigation";
-  }
-  if (props === 265259) {
-    source = TwistOfFateIcon;
-    alt = "Twist of Fate";
-  }
-  if (props === 214621) {
-    source = SchismIcon;
-    alt = "Schism";
-  }
-  if (props === 64129) {
-    source = BodyAndSoulIconDisc;
-    alt = "Body and Soul";
-  }
-  if (props === 193063) {
-    source = MasochismIcon;
-    alt = "Masochism";
-  }
-  if (props === 121536) {
-    source = AngelicFeatherIconDisc;
-    alt = "Angelic Feather";
-  }
-  if (props === 197045) {
-    source = ShieldDisciplineIcon;
-    alt = "Shield Discipline";
-  }
-  if (props === 123040) {
-    source = MindbenderIcon;
-    alt = "Mindbender";
-  }
-  if (props === 129250) {
-    source = PowerWordSolaceIcon;
-    alt = "Power Word: Solace";
-  }
-  if (props === 196704) {
-    source = PsychicVoiceIconDisc;
-    alt = "Psychic Voice";
-  }
-  if (props === 205367) {
-    source = DominantMindIcon;
-    alt = "Dominant Mind";
-  }
-  if (props === 204263) {
-    source = ShiningForceIconDisc;
-    alt = "Shining Force";
-  }
-  if (props === 280391) {
-    source = SinsOfTheManyIcon;
-    alt = "Sins of the Many";
-  }
-  if (props === 197419) {
-    source = ContritionIcon;
-    alt = "Contrition";
-  }
-  if (props === 314867) {
-    source = ShadowCovenantIcon;
-    alt = "Shadow Covenant";
-  }
-  if (props === 204197) {
-    source = PurgeTheWickedIcon;
-    alt = "Purge the Wicked";
-  }
-  if (props === 110744) {
-    source = DivineStarIconDisc;
-    alt = "Divine Star";
-  }
-  if (props === 120517) {
-    source = HaloIconDisc;
-    alt = "Halo";
-  }
-  if (props === 238063) {
-    source = LenienceIcon;
-    alt = "Lenience";
-  }
-  if (props === 109964) {
-    source = SpiritShellIcon;
-    alt = "Spirit Shell";
-  }
-  if (props === 246287) {
-    source = EvangelismIcon;
-    alt = "Evangelism";
-  }
-
-  // Priest - Holy
-
-  if (props === 193155) {
-    source = EnlightenmentIcon;
-    alt = "Enlightenment";
-  }
-  if (props === 200128) {
-    source = TrailOfLightIcon;
-    alt = "Trail of Light";
-  }
-  if (props === 341997) {
-    source = RenewedFaithIcon;
-    alt = "Renewed Faith";
-  }
-  if (props === 238100) {
-    source = AngelsMercyIcon;
-    alt = "Angel's Mercy";
-  }
-  if (props === 64129) {
-    source = BodyAndSoulIconHoly;
-    alt = "Body and Soul";
-  }
-  if (props === 121536) {
-    source = AngelicFeatherIconHoly;
-    alt = "Angelic Feather";
-  }
-  if (props === 238136) {
-    source = CosmicRippleIcon;
-    alt = "Cosmic Ripple";
-  }
-  if (props === 200209) {
-    source = GuardianAngelIcon;
-    alt = "Guardian Angel";
-  }
-  if (props === 196707) {
-    source = AfterlifeIcon;
-    alt = "Afterlife";
-  }
-  if (props === 196704) {
-    source = PsychicVoiceIconHoly;
-    alt = "Psychic Voice";
-  }
-  if (props === 200199) {
-    source = CensureIcon;
-    alt = "Censure";
-  }
-  if (props === 204263) {
-    source = ShiningForceIconHoly;
-    alt = "Shining Force";
-  }
-  if (props === 109186) {
-    source = SurgeOfLightIcon;
-    alt = "Surge of Light";
-  }
-  if (props === 32546) {
-    source = BindingHealIcon;
-    alt = "Binding Heal";
-  }
-  if (props === 321377) {
-    source = PrayerCircleIcon;
-    alt = "Prayer Circle";
-  }
-
-  if (props === 193157) {
-    source = BenedictionIcon;
-    alt = "Benediction";
-  }
-  if (props === 110744) {
-    source = DivineStarIconHoly;
-    alt = "Divine Star";
-  }
-  if (props === 120517) {
-    source = HaloIconHoly;
-    alt = "Halo";
-  }
-  if (props === 196985) {
-    source = LightOfTheNaaruIcon;
-    alt = "Light of the Naaru";
-  }
-  if (props === 200183) {
-    source = ApotheosisIcon;
-    alt = "Apotheosis";
-  }
-  if (props === 265202) {
-    source = HolyWordSalvationIcon;
-    alt = "Holy Word: Salvation";
-  }
-
-  return (
-    <a data-wowhead={"spell=" + props}>
-      <img
-        style={{
-          height: 30,
-          width: 30,
-          padding: 4,
-          verticalAlign: "middle",
-          borderRadius: "4px"
-        }}
-        src={source}
-        alt={alt}
-      />
-    </a>
-  );
-}
