+++ src/Modules/QEModules/CooldownPlanner/BasicComponents/CheckBox.js
