+++ src/Modules/QEModules/CooldownPlanner/BasicComponents/Switch.js
