+++ src/Modules/QEModules/CooldownPlanner/Chart/Chart.js
