+++ src/Modules/QEModules/CooldownPlanner/Chart/DTPSBarChart.js
