+++ src/Modules/QEModules/CooldownPlanner/Data/Data.js
@@ -424,56 +424,56 @@ export const bossAbilities = [
     ability: "Gluttonous Miasma",
     guid: 329298,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/GluttonousMiasmaIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/GluttonousMiasmaIcon.jpg"),
   },
   {
     bossID: 2383,
     ability: "Consume",
     guid: 334522,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/ConsumeIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/ConsumeIcon.jpg"),
   },
   {
     bossID: 2383,
     ability: "Expunge",
     guid: 329742,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/ExpungeIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/ExpungeIcon.jpg"),
   },
   {
     bossID: 2383,
     ability: "Obliterating Rift",
     guid: 329835,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/ObliteratingRiftIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/ObliteratingRiftIcon.jpg"),
   },
   {
     bossID: 2383,
     ability: "Volatile Ejection",
     guid: 334064,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/VolatileEjectionIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/VolatileEjectionIcon.jpg"),
   },
   {
     bossID: 2383,
     ability: "Desolate",
     guid: 329455,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/DesolateIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/DesolateIcon.jpg"),
   },
   {
     bossID: 2383,
     ability: "Hungering Strikes",
     guid: 332294,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/HungeringStrikesIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/HungeringStrikesIcon.jpg"),
   },
   {
     bossID: 2383,
     ability: "Overwhelm",
     guid: 329774,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/OverwhelmIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/OverwhelmIcon.jpg"),
   },
 
   // Lady Inerva Darkvein
@@ -482,175 +482,175 @@ export const bossAbilities = [
     ability: "Loose Anima",
     guid: 325184,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/LooseAnimaIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/LooseAnimaIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Container Breach",
     guid: 325225,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/ContainerBreachIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/ContainerBreachIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Expose Desires",
     guid: 341621,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/ExposeDesiresIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/ExposeDesiresIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Warped Desires",
     guid: 325382,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/WarpedDesiresIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/WarpedDesiresIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Hidden Desire",
     guid: 335322,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/HiddenDesireIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/HiddenDesireIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Expose Cognition",
     guid: 341623,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/ExposeCognitionIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/ExposeCognitionIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Shared Cognition",
     guid: 325908,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/SharedCognitionIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/SharedCognitionIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Expose Heart",
     guid: 341625,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/ExposeHeartIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/ExposeHeartIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Change of Heart",
     guid: 325384,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/ChangeOfHeartIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/ChangeOfHeartIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Bottled Anima",
     guid: 325769,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/BottledAnimaIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/BottledAnimaIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Unleashed Volatility",
     guid: 329618,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/UnleashedVolatilityIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/UnleashedVolatilityIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Lingering Anima",
     guid: 342281,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/LingeringAnimaIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/LingeringAnimaIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Shared Suffering",
     guid: 324983,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/SharedSufferingIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/SharedSufferingIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Indemnification",
     guid: 331527,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/IndemnificationIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/IndemnificationIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Sins and Suffering",
     guid: 342288,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/SinsAndSufferingIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/SinsAndSufferingIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Anima Web",
     guid: 326538,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/AnimaWebIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/AnimaWebIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Greater Sins and Suffering",
     guid: 342290,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/GreaterSinsAndSufferingIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/GreaterSinsAndSufferingIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Unleashed Shadow",
     guid: 341590,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/UnleashedShadowIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/UnleashedShadowIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Lightly Concentrated Anima",
     guid: 342320,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/LightlyConcentratedAnimaIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/LightlyConcentratedAnimaIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Concentrated Anima",
     guid: 342321,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/ConcentratedAnimaIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/ConcentratedAnimaIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Highly Concentrated Anima",
     guid: 342322,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/HighlyConcentratedAnimaIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/HighlyConcentratedAnimaIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Unconscionable Guilt",
     guid: 331573,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/UnconscionableGuiltIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/UnconscionableGuiltIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Condemn",
     guid: 331550,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/CondemnIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/CondemnIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Condemn",
     guid: 334017,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/CondemnIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/CondemnIcon.jpg"),
   },
   {
     bossID: 2406,
     ability: "Fragments of Shadow",
     guid: 325596,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/FragmentsOfShadowIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/FragmentsOfShadowIcon.jpg"),
   },
 
   // Shriekwing
@@ -659,105 +659,105 @@ export const bossAbilities = [
     ability: "Reverberating Scream",
     guid: 344112,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Shriekwing/Icons/ReverberatingScreamIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Shriekwing/Icons/ReverberatingScreamIcon.jpg"),
   },
   {
     bossID: 2398,
     ability: "Reverberating Scream",
     guid: 344114,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/Shriekwing/Icons/ReverberatingScreamIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Shriekwing/Icons/ReverberatingScreamIcon.jpg"),
   },
   {
     bossID: 2398,
     ability: "Deadly Descent",
     guid: 343021,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/Shriekwing/Icons/DeadlyDescentIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Shriekwing/Icons/DeadlyDescentIcon.jpg"),
   },
   {
     bossID: 2398,
     ability: "Echoing Sonar",
     guid: 343022,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/Shriekwing/Icons/EchoingSonarIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Shriekwing/Icons/EchoingSonarIcon.jpg"),
   },
   {
     bossID: 2398,
     ability: "Bloodcurdling Shriek",
     guid: 330712,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Shriekwing/Icons/BloodcurdlingShriekIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Shriekwing/Icons/BloodcurdlingShriekIcon.jpg"),
   },
   {
     bossID: 2398,
     ability: "Bloodcurdling Shriek",
     guid: 336005,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/Shriekwing/Icons/BloodcurdlingShriekIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Shriekwing/Icons/BloodcurdlingShriekIcon.jpg"),
   },
   {
     bossID: 2398,
     ability: "Sonar Shriek",
     guid: 343023,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Shriekwing/Icons/SonarShriekIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Shriekwing/Icons/SonarShriekIcon.jpg"),
   },
   {
     bossID: 2398,
     ability: "Blind Swipe",
     guid: 343005,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/Shriekwing/Icons/BlindSwipeIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Shriekwing/Icons/BlindSwipeIcon.jpg"),
   },
   {
     bossID: 2398,
     ability: "Sanguine Ichor",
     guid: 340324,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/Shriekwing/Icons/SanguineIchorIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Shriekwing/Icons/SanguineIchorIcon.jpg"),
   },
   {
     bossID: 2398,
     ability: "Descent",
     guid: 342923,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/Shriekwing/Icons/DescentIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Shriekwing/Icons/DescentIcon.jpg"),
   },
   {
     bossID: 2398,
     ability: "Bloodlight",
     guid: 343384,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Shriekwing/Icons/BloodlightIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Shriekwing/Icons/BloodlightIcon.jpg"),
   },
   {
     bossID: 2398,
     ability: "Exsanguinating Bite",
     guid: 328887,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/Shriekwing/Icons/ExsanguinatingBiteIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Shriekwing/Icons/ExsanguinatingBiteIcon.jpg"),
   },
   {
     bossID: 2398,
     ability: "Echoing Screech",
     guid: 342866,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/Shriekwing/Icons/EchoingScreechIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Shriekwing/Icons/EchoingScreechIcon.jpg"),
   },
   {
     bossID: 2398,
     ability: "Exsanguinated",
     guid: 328897,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/Shriekwing/Icons/ExsanguinatedIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Shriekwing/Icons/ExsanguinatedIcon.jpg"),
   },
   {
     bossID: 2398,
     ability: "Earsplitting Shriek",
     guid: 330711,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Shriekwing/Icons/EarsplittingShriekIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Shriekwing/Icons/EarsplittingShriekIcon.jpg"),
   },
 
   // Stone Born Generals
@@ -767,84 +767,84 @@ export const bossAbilities = [
     ability: "Wicked Laceration",
     guid: 333913,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/WickedLacerationIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/WickedLacerationIcon.jpg"),
   },
   {
     bossID: 2417,
     ability: "Heart Hemorrhage",
     guid: 334771,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/HeartHemorrhageIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/HeartHemorrhageIcon.jpg"),
   },
   {
     bossID: 2417,
     ability: "Pulverizing Meteor",
     guid: 339728,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/PulverizingMeteorIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/PulverizingMeteorIcon.jpg"),
   },
   {
     bossID: 2417,
     ability: "Wicked Blast",
     guid: 333716,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/WickedBlastIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/WickedBlastIcon.jpg"),
   },
   {
     bossID: 2417,
     ability: "Heart Rend",
     guid: 334765,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/HeartRendIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/HeartRendIcon.jpg"),
   },
   {
     bossID: 2417,
     ability: "Stone Spike",
     guid: 343063,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/StoneSpikeIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/StoneSpikeIcon.jpg"),
   },
   {
     bossID: 2417,
     ability: "Ravenous Feast",
     guid: 342733,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/RavenousFeastIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/RavenousFeastIcon.jpg"),
   },
   {
     bossID: 2417,
     ability: "Wicked Slaughter",
     guid: 342253,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/WickedSlaughterIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/WickedSlaughterIcon.jpg"),
   },
   {
     bossID: 2417,
     ability: "Serrated Swipe",
     guid: 334929,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/SerratedSwipeIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/SerratedSwipeIcon.jpg"),
   },
   {
     bossID: 2417,
     ability: "Crystalline Burst",
     guid: 339693,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/CrystallineBurstIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/CrystallineBurstIcon.jpg"),
   },
   {
     bossID: 2417,
     ability: "Crystalize",
     guid: 339690,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/CrystalizeIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/CrystalizeIcon.jpg"),
   },
   {
     bossID: 2417,
     ability: "Ravenous Swipe",
     guid: 342740,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/RavenousSwipeIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/RavenousSwipeIcon.jpg"),
   },
 
   // Sun King's Salvation
@@ -853,119 +853,119 @@ export const bossAbilities = [
     ability: "Smoldering Remnants",
     guid: 328579,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/SmolderingRemnantsIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/SmolderingRemnantsIcon.jpg"),
   },
   {
     bossID: 2402,
     ability: "Smoldering Plumage",
     guid: 328659,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/SmolderingPlumageIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/SmolderingPlumageIcon.jpg"),
   },
   {
     bossID: 2402,
     ability: "Fiery Strike",
     guid: 326455,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/FieryStrikeIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/FieryStrikeIcon.jpg"),
   },
   {
     bossID: 2402,
     ability: "Burning Remnants",
     guid: 326456,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/BurningRemnantsIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/BurningRemnantsIcon.jpg"),
   },
   {
     bossID: 2402,
     ability: "Ember Blast",
     guid: 325877,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/EmberBlastIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/EmberBlastIcon.jpg"),
   },
   {
     bossID: 2402,
     ability: "Lingering Embers",
     guid: 326430,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/LingeringEmbersIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/LingeringEmbersIcon.jpg"),
   },
   {
     bossID: 2402,
     ability: "Blazing Surge",
     guid: 329518,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/BlazingSurgeIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/BlazingSurgeIcon.jpg"),
   },
   {
     bossID: 2402,
     ability: "Vanquishing Strike",
     guid: 325440,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/VanquishingStrikeIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/VanquishingStrikeIcon.jpg"),
   },
   {
     bossID: 2402,
     ability: "Vanquished",
     guid: 325442,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/VanquishedIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/VanquishedIcon.jpg"),
   },
   {
     bossID: 2402,
     ability: "Concussive Smash",
     guid: 325506,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/ConcussiveSmashIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/ConcussiveSmashIcon.jpg"),
   },
   {
     bossID: 2402,
     ability: "Crimson Flurry",
     guid: 341473,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/CrimsonFlurryIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/CrimsonFlurryIcon.jpg"),
   },
   {
     bossID: 2402,
     ability: "Vulgar Brand",
     guid: 333002,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/VulgarBrandIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/VulgarBrandIcon.jpg"),
   },
   {
     bossID: 2402,
     ability: "Scornful Blast",
     guid: 325590,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/ScornfulBlastIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/ScornfulBlastIcon.jpg"),
   },
   {
     bossID: 2402,
     ability: "Shattering Ruby",
     guid: 335540,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/ShatteringRubyIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/ShatteringRubyIcon.jpg"),
   },
   {
     bossID: 2402,
     ability: "Fragmentation",
     guid: 336398,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/FragmentationIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/FragmentationIcon.jpg"),
   },
   {
     bossID: 2402,
     ability: "Greater Castigation",
     guid: 328890,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/GreaterCastigationIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/GreaterCastigationIcon.jpg"),
   },
   {
     bossID: 2402,
     ability: "Soul Infusion",
     guid: 339232,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/SoulInfusionIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/SoulInfusionIcon.jpg"),
   },
 
   // Huntsman Altimor
@@ -975,105 +975,105 @@ export const bossAbilities = [
     ability: "Sinseeker",
     guid: 335114,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/SinseekerIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/SinseekerIcon.jpg"),
   },
   {
     bossID: 2418,
     ability: "Vicious Bolt",
     guid: 338615,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/ViciousBoltIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/ViciousBoltIcon.jpg"),
   },
   {
     bossID: 2418,
     ability: "Pierce Soul",
     guid: 338609,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/PierceSoulIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/PierceSoulIcon.jpg"),
   },
   {
     bossID: 2418,
     ability: "Shatter Shot",
     guid: 338593,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/ShatterShotIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/ShatterShotIcon.jpg"),
   },
   {
     bossID: 2418,
     ability: "Spreadshot",
     guid: 334404,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/SpreadshotIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/SpreadshotIcon.jpg"),
   },
   {
     bossID: 2418,
     ability: "Vicious Lunge",
     guid: 334939,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/ViciousLungeIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/ViciousLungeIcon.jpg"),
   },
   {
     bossID: 2418,
     ability: "Vicious Wound",
     guid: 334960,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/ViciousWoundIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/ViciousWoundIcon.jpg"),
   },
   {
     bossID: 2418,
     ability: "Jagged Claws",
     guid: 334971,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/JaggedClawsIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/JaggedClawsIcon.jpg"),
   },
   {
     bossID: 2418,
     ability: "Deathly Roar",
     guid: 334708,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/DeathlyRoarIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/DeathlyRoarIcon.jpg"),
   },
   {
     bossID: 2418,
     ability: "Rip Soul",
     guid: 334797,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/RipSoulIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/RipSoulIcon.jpg"),
   },
   {
     bossID: 2418,
     ability: "Unstable Soul",
     guid: 339639,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/UnstableSoulIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/UnstableSoulIcon.jpg"),
   },
   {
     bossID: 2418,
     ability: "Crushing Stone",
     guid: 334860,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/CrushingStoneIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/CrushingStoneIcon.jpg"),
   },
   {
     bossID: 2418,
     ability: "Petrifying Howl",
     guid: 334852,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/PetrifyingHowlIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/PetrifyingHowlIcon.jpg"),
   },
   {
     bossID: 2418,
     ability: "Stone Shards",
     guid: 334893,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/StoneShardsIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/StoneShardsIcon.jpg"),
   },
   {
     bossID: 2418,
     ability: "Shoot",
     guid: 335016,
     cooldownPlannerActive: false,
-    icon: require("./../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/ShootIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/ShootIcon.jpg"),
   },
 
   // Xymox
@@ -1082,70 +1082,70 @@ export const bossAbilities = [
     ability: "Dimensional Tear",
     guid: 328437,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/DimensionalTearIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/DimensionalTearIcon.jpg"),
   },
   {
     bossID: 2405,
     ability: "Glyph of Destruction",
     guid: 325361,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/GlyphOfDestructionIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/GlyphOfDestructionIcon.jpg"),
   },
   {
     bossID: 2405,
     ability: "Stasis Trap",
     guid: 326271,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/StasisTrapIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/StasisTrapIcon.jpg"),
   },
   {
     bossID: 2405,
     ability: "Rift Blast",
     guid: 329458,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/RiftBlastIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/RiftBlastIcon.jpg"),
   },
   {
     bossID: 2405,
     ability: "Hyperlight Spark",
     guid: 325399,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/HyperlightSparkIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/HyperlightSparkIcon.jpg"),
   },
   {
     bossID: 2405,
     ability: "Root of Extinction",
     guid: 329770,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/RootOfExtinctionIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/RootOfExtinctionIcon.jpg"),
   },
   {
     bossID: 2405,
     ability: "Withering Touch",
     guid: 340860,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/WitheringTouchIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/WitheringTouchIcon.jpg"),
   },
   {
     bossID: 2405,
     ability: "Edge of Annihilation",
     guid: 328880,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/EdgeOfAnnihilationIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/EdgeOfAnnihilationIcon.jpg"),
   },
   {
     bossID: 2405,
     ability: "Aura of Dread",
     guid: 340870,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/AuraOfDreadIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/AuraOfDreadIcon.jpg"),
   },
   {
     bossID: 2405,
     ability: "Soul Singe",
     guid: 340842,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/SoulSingeIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/SoulSingeIcon.jpg"),
   },
 
   // Council of Blood
@@ -1154,63 +1154,63 @@ export const bossAbilities = [
     ability: "Oppressive Atmosphere",
     guid: 334909,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/OppressiveAtmosphereIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/OppressiveAtmosphereIcon.jpg"),
   },
   {
     bossID: 2412,
     ability: "Tactical Advance",
     guid: 328334,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/TacticalAdvanceIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/TacticalAdvanceIcon.jpg"),
   },
   {
     bossID: 2412,
     ability: "Unstoppable Charge",
     guid: 334948,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/UnstoppableChargeIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/UnstoppableChargeIcon.jpg"),
   },
   {
     bossID: 2412,
     ability: "Bolt of Power",
     guid: 337110,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/BoltOfPowerIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/BoltOfPowerIcon.jpg"),
   },
   {
     bossID: 2412,
     ability: "Anima Fountain",
     guid: 327475,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/AnimaFountainIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/AnimaFountainIcon.jpg"),
   },
   {
     bossID: 2412,
     ability: "Scarlet Letter",
     guid: 331708,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/ScarletLetterIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/ScarletLetterIcon.jpg"),
   },
   {
     bossID: 2412,
     ability: "Throw Food",
     guid: 330968,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/ThrowFoodIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/ThrowFoodIcon.jpg"),
   },
   {
     bossID: 2412,
     ability: "Evasive Lunge",
     guid: 327503,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/EvasiveLungeIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/EvasiveLungeIcon.jpg"),
   },
   {
     bossID: 2412,
     ability: "Waltz of Blood",
     guid: 327619,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/WaltzOfBloodIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/WaltzOfBloodIcon.jpg"),
   },
   // From the adds
   {
@@ -1218,14 +1218,14 @@ export const bossAbilities = [
     ability: "Waltz of Blood",
     guid: 342852,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/WaltzOfBlood2Icon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/WaltzOfBlood2Icon.jpg"),
   },
   {
     bossID: 2412,
     ability: "Dark Recital",
     guid: 331634,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/DarkRecitalIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/DarkRecitalIcon.jpg"),
   },
 
   // sludgefist
@@ -1234,105 +1234,105 @@ export const bossAbilities = [
     ability: "Heedless Charge",
     guid: 331212,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Sludgefist/Icons/HeedlessChargeIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Sludgefist/Icons/HeedlessChargeIcon.jpg"),
   },
   {
     bossID: 2399,
     ability: "Destructive Impact",
     guid: 332969,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Sludgefist/Icons/DestructiveImpactIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Sludgefist/Icons/DestructiveImpactIcon.jpg"),
   },
   {
     bossID: 2399,
     ability: "Crumbling Foundation",
     guid: 332443,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Sludgefist/Icons/CrumblingFoundationIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Sludgefist/Icons/CrumblingFoundationIcon.jpg"),
   },
   {
     bossID: 2399,
     ability: "Collapsing Foundation",
     guid: 332197,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Sludgefist/Icons/CollapsingFoundationIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Sludgefist/Icons/CollapsingFoundationIcon.jpg"),
   },
   {
     bossID: 2399,
     ability: "Fractured Boulder",
     guid: 341102,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Sludgefist/Icons/FracturedBoulderIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Sludgefist/Icons/FracturedBoulderIcon.jpg"),
   },
   {
     bossID: 2399,
     ability: "Fractured Impact",
     guid: 341482,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Sludgefist/Icons/FracturedImpactIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Sludgefist/Icons/FracturedImpactIcon.jpg"),
   },
   {
     bossID: 2399,
     ability: "Chain Link",
     guid: 335300,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Sludgefist/Icons/ChainLinkIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Sludgefist/Icons/ChainLinkIcon.jpg"),
   },
   {
     bossID: 2399,
     ability: "Shattering Chain",
     guid: 335295,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Sludgefist/Icons/ShatteringChainIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Sludgefist/Icons/ShatteringChainIcon.jpg"),
   },
   {
     bossID: 2399,
     ability: "Destructive Stomp",
     guid: 332318,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Sludgefist/Icons/DestructiveStompIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Sludgefist/Icons/DestructiveStompIcon.jpg"),
   },
   {
     bossID: 2399,
     ability: "Seismic Shift",
     guid: 340803,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Sludgefist/Icons/SeismicShiftIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Sludgefist/Icons/SeismicShiftIcon.jpg"),
   },
   {
     bossID: 2399,
     ability: "Falling Rubble",
     guid: 332572,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Sludgefist/Icons/FallingRubbleIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Sludgefist/Icons/FallingRubbleIcon.jpg"),
   },
   {
     bossID: 2399,
     ability: "Stonequake",
     guid: 335361,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Sludgefist/Icons/StonequakeIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Sludgefist/Icons/StonequakeIcon.jpg"),
   },
   {
     bossID: 2399,
     ability: "Colossal Roar",
     guid: 332687,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Sludgefist/Icons/ColossalRoarIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Sludgefist/Icons/ColossalRoarIcon.jpg"),
   },
   {
     bossID: 2399,
     ability: "Chain Slam",
     guid: 335354,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Sludgefist/Icons/ChainSlamIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Sludgefist/Icons/ChainSlamIcon.jpg"),
   },
   {
     bossID: 2399,
     ability: "Chain Bleed",
     guid: 339189,
     cooldownPlannerActive: true,
-    icon: require("./../../../Images/Bosses/CastleNathria/Sludgefist/Icons/ChainBleedIcon.jpg"),
+    icon: require("../../../../Images/Bosses/CastleNathria/Sludgefist/Icons/ChainBleedIcon.jpg"),
   },
 ];
