+++ src/Modules/QEModules/CooldownPlanner/Functions/ClassColourFunctions.js
