+++ src/Modules/QEModules/CooldownPlanner/Functions/Functions.js
