+++ src/Modules/QEModules/CooldownPlanner/Functions/IconFunctions/AbilityIcons.js
@@ -1,37 +1,38 @@
 import React from "react";
-import AuraMasteryIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/AuraMastery.jpg";
-import AvengingWrathIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/AvengingWrath.jpg";
-import DivineHymnIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/DivineHymn.jpg";
-import EvangelismIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Evangelism.jpg";
-import HealingTideTotemIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/HealingTideTotem.jpg";
-import PowerWordBarrierIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/PowerWordBarrier.jpg";
-import RevivalIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Revival.jpg";
-import SalvationIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Salvation.jpg";
-import SpiritLinkTotemIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/SpiritLinkTotem.jpg";
-import TranquilityIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Tranquility.jpg";
-import TreeofLifeIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/TreeofLife.jpg";
-import DivineTollIcon from "../../../../Images/Classes/Paladin/CovenantAbilities/DivineToll.jpg";
-import AshenHallowIcon from "../../../../Images/Classes/Paladin/CovenantAbilities/AshenHallow.jpg";
-import VanquishersHammerIcon from "../../../../Images/Classes/Paladin/CovenantAbilities/VanquishersHammer.jpg";
-import BlessingOfSeasonsIcon from "../../../../Images/Classes/Paladin/CovenantAbilities/BlessingOfSeasons.jpg";
-import KindredSpiritsIcon from "../../../../Images/Classes/Druid/CovenantAbilities/KindredSpirits.jpg";
-import RavenousFrenzyIcon from "../../../../Images/Classes/Druid/CovenantAbilities/RavenousFrenzy.jpg";
-import AdaptiveSwarmIcon from "../../../../Images/Classes/Druid/CovenantAbilities/AdaptiveSwarm.jpg";
-import ConvokeTheSpiritsIcon from "../../../../Images/Classes/Druid/CovenantAbilities/ConvokeTheSpirits.jpg";
-import BoonoftheAscendedIcon from "../../../../Images/Classes/Priest/CovenantAbilities/BoonoftheAscended.jpg";
-import MindgamesIcon from "../../../../Images/Classes/Priest/CovenantAbilities/Mindgames.jpg";
-import UnholyNovaIcon from "../../../../Images/Classes/Priest/CovenantAbilities/UnholyNova.jpg";
-import FaeBlessingsIcon from "../../../../Images/Classes/Priest/CovenantAbilities/FaeBlessings.jpg";
-import VesperTotemIcon from "../../../../Images/Classes/Shaman/CovenantAbilities/VesperTotem.jpg";
-import ChainHarvestIcon from "../../../../Images/Classes/Shaman/CovenantAbilities/ChainHarvest.jpg";
-import PrimordialWaveIcon from "../../../../Images/Classes/Shaman/CovenantAbilities/PrimordialWave.jpg";
-import FaeTransfusionIcon from "../../../../Images/Classes/Shaman/CovenantAbilities/FaeTransfusion.jpg";
-import WeaponsofOrderIcon from "../../../../Images/Classes/Monk/CovenantAbilities/WeaponsofOrder.jpg";
-import FallenOrderIcon from "../../../../Images/Classes/Monk/CovenantAbilities/FallenOrder.jpg";
-import BonedustBrewIcon from "../../../../Images/Classes/Monk/CovenantAbilities/BonedustBrew.jpg";
-import FaelineStompIcon from "../../../../Images/Classes/Monk/CovenantAbilities/FaelineStomp.jpg";
-import DarknessIcon from "../../../../Images/Classes/DemonHunter/DemonHunterIcon.jpg";
-import RallyingCryIcon from "../../../../Images/Classes/Warrior/RallyingCryIcon.jpg";
+import AuraMasteryIcon from "../../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/AuraMastery.jpg";
+import AvengingWrathIcon from "../../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/AvengingWrath.jpg";
+import DivineHymnIcon from "../../../../../Images/Classes/Priest/Specialisation/Holy/Icons/DivineHymn.jpg";
+import EvangelismIcon from "../../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Evangelism.jpg";
+import HealingTideTotemIcon from "../../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/HealingTideTotem.jpg";
+import PowerWordBarrierIcon from "../../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/PowerWordBarrier.jpg";
+import RevivalIcon from "../../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Revival.jpg";
+import SalvationIcon from "../../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Salvation.jpg";
+import SpiritLinkTotemIcon from "../../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/SpiritLinkTotem.jpg";
+import TranquilityIcon from "../../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Tranquility.jpg";
+import TreeofLifeIcon from "../../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/TreeofLife.jpg";
+import DivineTollIcon from "../../../../../Images/Classes/Paladin/CovenantAbilities/DivineToll.jpg";
+import AshenHallowIcon from "../../../../../Images/Classes/Paladin/CovenantAbilities/AshenHallow.jpg";
+import VanquishersHammerIcon from "../../../../../Images/Classes/Paladin/CovenantAbilities/VanquishersHammer.jpg";
+import BlessingOfSeasonsIcon from "../../../../../Images/Classes/Paladin/CovenantAbilities/BlessingOfSeasons.jpg";
+import KindredSpiritsIcon from "../../../../../Images/Classes/Druid/CovenantAbilities/KindredSpirits.jpg";
+import RavenousFrenzyIcon from "../../../../../Images/Classes/Druid/CovenantAbilities/RavenousFrenzy.jpg";
+import AdaptiveSwarmIcon from "../../../../../Images/Classes/Druid/CovenantAbilities/AdaptiveSwarm.jpg";
+import ConvokeTheSpiritsIcon from "../../../../../Images/Classes/Druid/CovenantAbilities/ConvokeTheSpirits.jpg";
+import BoonoftheAscendedIcon from "../../../../../Images/Classes/Priest/CovenantAbilities/BoonoftheAscended.jpg";
+import MindgamesIcon from "../../../../../Images/Classes/Priest/CovenantAbilities/Mindgames.jpg";
+import UnholyNovaIcon from "../../../../../Images/Classes/Priest/CovenantAbilities/UnholyNova.jpg";
+import FaeBlessingsIcon from "../../../../../Images/Classes/Priest/CovenantAbilities/FaeBlessings.jpg";
+import VesperTotemIcon from "../../../../../Images/Classes/Shaman/CovenantAbilities/VesperTotem.jpg";
+import ChainHarvestIcon from "../../../../../Images/Classes/Shaman/CovenantAbilities/ChainHarvest.jpg";
+import PrimordialWaveIcon from "../../../../../Images/Classes/Shaman/CovenantAbilities/PrimordialWave.jpg";
+import FaeTransfusionIcon from "../../../../../Images/Classes/Shaman/CovenantAbilities/FaeTransfusion.jpg";
+import WeaponsofOrderIcon from "../../../../../Images/Classes/Monk/CovenantAbilities/WeaponsofOrder.jpg";
+import FallenOrderIcon from "../../../../../Images/Classes/Monk/CovenantAbilities/FallenOrder.jpg";
+import BonedustBrewIcon from "../../../../../Images/Classes/Monk/CovenantAbilities/BonedustBrew.jpg";
+import FaelineStompIcon from "../../../../../Images/Classes/Monk/CovenantAbilities/FaelineStomp.jpg";
+import DarknessIcon from "../../../../../Images/Classes/DemonHunter/DemonHunterIcon.jpg";
+import RallyingCryIcon from "../../../../../Images/Classes/Warrior/RallyingCryIcon.jpg";
+
 
 // Should probably work this off GUID
