+++ src/Modules/QEModules/CooldownPlanner/Functions/IconFunctions/BossAbilityIcons.js
