+++ src/Modules/QEModules/CooldownPlanner/Functions/IconFunctions/BossIcons.js
@@ -1,27 +1,27 @@
 import React from "react";
-import MautIcon from "../../../../Images/Bosses/Nya'lotha/Maut.jpg";
-import XaneshIcon from "../../../../Images/Bosses/Nya'lotha/Xanesh.jpg";
-import WrathionIcon from "../../../../Images/Bosses/Nya'lotha/Wrathion.jpg";
-import RadenIcon from "../../../../Images/Bosses/Nya'lotha/Raden.jpg";
-import HivemindIcon from "../../../../Images/Bosses/Nya'lotha/Hivemind.jpg";
-import SkitraIcon from "../../../../Images/Bosses/Nya'lotha/Skitra.jpg";
-import ShadharIcon from "../../../../Images/Bosses/Nya'lotha/Shadhar.jpg";
-import VexionaIcon from "../../../../Images/Bosses/Nya'lotha/Vexiona.jpg";
-import CarapaceIcon from "../../../../Images/Bosses/Nya'lotha/Carapace.jpg";
-import DrestagathIcon from "../../../../Images/Bosses/Nya'lotha/Drestagath.jpg";
-import NzothIcon from "../../../../Images/Bosses/Nya'lotha/Nzoth.jpg";
-import IlgynothIcon from "../../../../Images/Bosses/Nya'lotha/Ilgynoth.jpg";
+import MautIcon from "../../../../../Images/Bosses/Nya'lotha/Maut.jpg";
+import XaneshIcon from "../../../../../Images/Bosses/Nya'lotha/Xanesh.jpg";
+import WrathionIcon from "../../../../../Images/Bosses/Nya'lotha/Wrathion.jpg";
+import RadenIcon from "../../../../../Images/Bosses/Nya'lotha/Raden.jpg";
+import HivemindIcon from "../../../../../Images/Bosses/Nya'lotha/Hivemind.jpg";
+import SkitraIcon from "../../../../../Images/Bosses/Nya'lotha/Skitra.jpg";
+import ShadharIcon from "../../../../../Images/Bosses/Nya'lotha/Shadhar.jpg";
+import VexionaIcon from "../../../../../Images/Bosses/Nya'lotha/Vexiona.jpg";
+import CarapaceIcon from "../../../../../Images/Bosses/Nya'lotha/Carapace.jpg";
+import DrestagathIcon from "../../../../../Images/Bosses/Nya'lotha/Drestagath.jpg";
+import NzothIcon from "../../../../../Images/Bosses/Nya'lotha/Nzoth.jpg";
+import IlgynothIcon from "../../../../../Images/Bosses/Nya'lotha/Ilgynoth.jpg";
 
-import ArtificerIcon from "../../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/ArtificerIcon.jpg";
-import CouncilIcon from "../../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/TheCouncilOfBloodIcon.jpg";
-import HungeringDestroyerIcon from "../../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/HungeringDestroyerIcon.jpg";
-import HuntsmanAltimorIcon from "../../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/HuntsmanAltimorIcon.jpg";
-import KaelthasSunstriderIcon from "../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/KaelthasSunstriderIcon.jpg";
-import LadyInervaDarkveinIcon from "../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/LadyInervaDarkveinIcons.jpg";
-import ShriekwingIcon from "../../../../Images/Bosses/CastleNathria/Shriekwing/Icons/ShriekwingIcon.jpg";
-import SireDenathriusIcon from "../../../../Images/Bosses/CastleNathria/SireDenathrius/Icons/SireDenathriusIcon.jpg";
-import SludgefistIcon from "../../../../Images/Bosses/CastleNathria/Sludgefist/Icons/SludgefistIcon.jpg";
-import StonebornGeneralsIcon from "../../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/StonebornGeneralsIcons.jpg";
+import ArtificerIcon from "../../../../../Images/Bosses/CastleNathria/ArtificerXymox/Icons/ArtificerIcon.jpg";
+import CouncilIcon from "../../../../../Images/Bosses/CastleNathria/TheCouncilOfBlood/Icons/TheCouncilOfBloodIcon.jpg";
+import HungeringDestroyerIcon from "../../../../../Images/Bosses/CastleNathria/HungeringDestroyer/Icons/HungeringDestroyerIcon.jpg";
+import HuntsmanAltimorIcon from "../../../../../Images/Bosses/CastleNathria/HuntsmanAltimor/Icons/HuntsmanAltimorIcon.jpg";
+import KaelthasSunstriderIcon from "../../../../../Images/Bosses/CastleNathria/SunKingsSalvation/Icons/KaelthasSunstriderIcon.jpg";
+import LadyInervaDarkveinIcon from "../../../../../Images/Bosses/CastleNathria/LadyInervaDarkvein/Icons/LadyInervaDarkveinIcons.jpg";
+import ShriekwingIcon from "../../../../../Images/Bosses/CastleNathria/Shriekwing/Icons/ShriekwingIcon.jpg";
+import SireDenathriusIcon from "../../../../../Images/Bosses/CastleNathria/SireDenathrius/Icons/SireDenathriusIcon.jpg";
+import SludgefistIcon from "../../../../../Images/Bosses/CastleNathria/Sludgefist/Icons/SludgefistIcon.jpg";
+import StonebornGeneralsIcon from "../../../../../Images/Bosses/CastleNathria/StoneLegionGenerals/Icons/StonebornGeneralsIcons.jpg";
 
 export default function bossIcons(props) {
   let source = "";
