+++ src/Modules/QEModules/CooldownPlanner/Tables/ClassCooldownMenuItems.js
@@ -1,19 +1,20 @@
 import React from "react";
 import { MenuItem } from "@material-ui/core";
-import AuraMasteryIcon from "../../../Images/Classes/Paladin/Specialisation/Holy/Icons/AuraMastery.jpg";
-import AvengingWrathIcon from "../../../Images/Classes/Paladin/Specialisation/Holy/Icons/AvengingWrath.jpg";
-import DivineHymnIcon from "../../../Images/Classes/Priest/Specialisation/Holy/Icons/DivineHymn.jpg";
-import EvangelismIcon from "../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Evangelism.jpg";
-import HealingTideTotemIcon from "../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/HealingTideTotem.jpg";
-import PowerWordBarrierIcon from "../../../Images/Classes/Priest/Specialisation/Discipline/Icons/PowerWordBarrier.jpg";
-import RevivalIcon from "../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Revival.jpg";
-import SalvationIcon from "../../../Images/Classes/Priest/Specialisation/Holy/Icons/Salvation.jpg";
-import SpiritLinkTotemIcon from "../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/SpiritLinkTotem.jpg";
-import TranquilityIcon from "../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Tranquility.jpg";
-import TreeofLifeIcon from "../../../Images/Classes/Druid/Specialisation/Restoration/Icons/TreeofLife.jpg";
-import DarknessIcon from "../../../Images/Classes/DemonHunter/DemonHunterIcon.jpg"
-import RallyingCryIcon from "../../../Images/Classes/Warrior/RallyingCryIcon.jpg"
+import AuraMasteryIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/AuraMastery.jpg";
+import AvengingWrathIcon from "../../../../Images/Classes/Paladin/Specialisation/Holy/Icons/AvengingWrath.jpg";
+import DivineHymnIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/DivineHymn.jpg";
+import EvangelismIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/Evangelism.jpg";
+import HealingTideTotemIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/HealingTideTotem.jpg";
+import PowerWordBarrierIcon from "../../../../Images/Classes/Priest/Specialisation/Discipline/Icons/PowerWordBarrier.jpg";
+import RevivalIcon from "../../../../Images/Classes/Monk/Specialisation/Mistweaver/Icons/Revival.jpg";
+import SalvationIcon from "../../../../Images/Classes/Priest/Specialisation/Holy/Icons/Salvation.jpg";
+import SpiritLinkTotemIcon from "../../../../Images/Classes/Shaman/Specialisation/Restoration/Icons/SpiritLinkTotem.jpg";
+import TranquilityIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/Tranquility.jpg";
+import TreeofLifeIcon from "../../../../Images/Classes/Druid/Specialisation/Restoration/Icons/TreeofLife.jpg";
+import DarknessIcon from "../../../../Images/Classes/DemonHunter/DemonHunterIcon.jpg"
+import RallyingCryIcon from "../../../../Images/Classes/Warrior/RallyingCryIcon.jpg"
 import Divider from "@material-ui/core/Divider";
+
 
 export default function ClassCooldownMenuItems(props) {
   // Paladin Cooldowns
