+++ src/Modules/QEModules/CooldownPlanner/Tables/CooldownPlanner.js
