+++ src/Modules/QEModules/CooldownPlanner/Tables/DtpsTable.js
