+++ src/Modules/QEModules/CooldownPlanner/Tables/HealerInfoTable.js
