+++ src/Modules/QEModules/CooldownPlanner/Tables/TableLocalization.js
