+++ src/Modules/QEModules/CooldownPlanner/TestFiles/LegendaryObject2.js
